+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - Added Deep Copy option to the Copy Object Data node.
 - Added Decompose Text node.
 - Added Object Color Output node.
+- Added GP Strokes From Splines node.
+- Added Splines From GP Strokes node.
 
 ### Fixed